% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pct_impervious.R
\name{download_or_check_impervious}
\alias{download_or_check_impervious}
\title{Download or check existence of impervious surface raster data}
\usage{
download_or_check_impervious(edition = "Annual_NLCD_FctImp_2024_CU_C1V1.tif")
}
\arguments{
\item{edition}{Character string specifying the NLCD data edition to use. Defaults to
"Annual_NLCD_FctImp_2024_CU_C1V1.tif".}
}
\value{
Function produces no output if ready to proceed; produces messaging if data is not available to the user.
}
\description{
This function manages the download and organization of NLCD impervious surface raster data.
It creates the necessary directory structure and provides guidance if the data is missing.
}
\details{
\enumerate{
\item Data should be downloaded from the \href{https://www.mrlc.gov/downloads/sciweb1/shared/mrlc/data-bundles/Annual_NLCD_FctImp_2024_CU_C1V1.zip}{USGS NLCD}.
Data files are quite large (~1GB).
\item Next, the zip file should be unpacked and the .tif file should be placed in the \code{urbanr_data} directory inside the working directory.
}

Alternatively, you can download a different snapshot from the NLCD
\href{https://www.mrlc.gov/data?f\%5B0\%5D=category\%3AFractional\%20Impervious\%20Surface}{here} and change the
\code{edition} argument to that .tif filename. File should still be placed in the proper
directory.

In the event that the data is no longer available from NLCD, the 2024 snapshot
can also be found on \href{https://figshare.com/articles/dataset/urbanr_data_Annual_NLCD_FctImp_2024/29549666?file=56194733}{Figshare}.
}
\examples{
download_or_check_impervious()
}
